[REFACTOR > MAIN] refactored UserService/Repository

+++ server/src/Modules/shared/database/Repositories/User/statistics.repository.ts
@@ -11,9 +11,9 @@ export class StatisticsRepository {
 
     /**
      * @description - Queries the database for all statistics
-     * @returns {Promise<SelectUserStatistics[] | [] | Error>} - Returns all statistics or an empty array if no statistics are found and an error if an error occurs
+     * @returns {Promise<SelectUserStatistics[] | Error>} - Returns all statistics or an empty array if no statistics are found and an error if an error occurs
      */
-    async queryAllStatistics(): Promise<SelectUserStatistics[] | [] | Error> {
+    async findAll(): Promise<SelectUserStatistics[] | Error> {
         try {
             return await this.db.select().from(schema.usersStatisticsTable);
         } catch (error) {
@@ -24,9 +24,9 @@ export class StatisticsRepository {
     /**
      * @description - Queries the database for a statistics by id
      * @param id - The id of the statistics
-     * @returns {Promise<SelectUserStatistics | [] | Error>} - Returns a statistics or an empty array if no statistics are found and an error if an error occurs
+     * @returns {Promise<SelectUserStatistics | Error>} - Returns a statistics or an empty array if no statistics are found and an error if an error occurs
      */
-    async queryStatisticsById(id: string): Promise<SelectUserStatistics | [] | Error> {
+    async findByID(id: string): Promise<SelectUserStatistics | Error> {
         try {
             return await this.db.query.usersStatisticsTable.findFirst({
                 where: eq(schema.usersStatisticsTable.user_id, id)
@@ -39,15 +39,15 @@ export class StatisticsRepository {
     /**
      * @description - Inserts a new statistics into the database
      * @param {string} id - The id of the user
-     * @returns {Promise<string | [] | Error>} - Returns the id of the user or null if an error occurs
+     * @returns {Promise<string | Error>} - Returns the id of the user or null if an error occurs
      */
-    async insertDefaultTable(id: string): Promise<string | [] | Error> {
+    async insertDefaultTable(id: string): Promise<string | Error> {
         try {
             const statistics = await this.db
                 .insert(schema.usersStatisticsTable)
                 .values({ user_id: id })
                 .returning({ user_id: schema.usersStatisticsTable.user_id });
-            return statistics ? statistics[0].user_id : [];
+            return statistics[0].user_id;
         } catch (error) {
             return error;
         }
@@ -57,16 +57,16 @@ export class StatisticsRepository {
      * @description - Updates a statistics by id
      * @param id - The id of the statistics
      * @param body - The body of the request
-     * @returns {Promise<string | [] | Error>} - Returns the userID or an empty array if no statistics are found and an error if an error occurs
+     * @returns {Promise<string | Error>} - Returns the userID or an empty array if no statistics are found and an error if an error occurs
      */
-    async updateStatistics(id: string, body: any): Promise<string | [] | Error> {
+    async updateOne(id: string, body: any): Promise<string | Error> {
         try {
             const statistics = await this.db
                 .update(schema.usersStatisticsTable)
                 .set(body)
                 .where(eq(schema.usersStatisticsTable.user_id, id))
                 .returning({ user_id: schema.usersStatisticsTable.user_id });
-            return statistics ? statistics[0].user_id : [];
+            return statistics[0].user_id; 
         } catch (error) {
             return error;
         }
@@ -77,10 +77,10 @@ export class StatisticsRepository {
      * @param id - The id of the statistics
      * @returns {Promise<string | [] | Error>} - Returns the id of the deleted statistics or null if an error occurs
      */
-    async deleteStatisticsById(id: string): Promise<string | [] | Error> {
+    async deleteOne(id: string): Promise<string | Error> {
         try {
-            await this.db.delete(schema.usersStatisticsTable).where(eq(schema.usersStatisticsTable.user_id, id));
-            return id;
+            const deleted = await this.db.delete(schema.usersStatisticsTable).where(eq(schema.usersStatisticsTable.user_id, id)).returning({ user_id: schema.usersStatisticsTable.user_id });
+            return deleted[0].user_id;
         } catch (error) {
             return error;
         }
